Muevo logica de login a clase Request
