+++ SiSMoS/app/src/main/java/com/jk/sismos/main/data/remote/Request.java
@@ -1,6 +1,7 @@
 package com.jk.sismos.main.data.remote;
 
 import android.util.Log;
+import android.widget.Toast;
 
 import com.jk.sismos.main.data.model.UserPost;
 
@@ -18,7 +19,26 @@ public class Request {
     public Request() {
         mAPIService = ApiUtils.getAPIService();
     }
+    public void sendLogin(String email, String password){
+        mAPIService.login(ENV, email, password, COMMISION, GROUP).enqueue(new Callback<UserPost>() {
+            @Override
+            public void onResponse(Call<UserPost> call, Response<UserPost> response) {
+                if (response.isSuccessful()) {
+                    showResponse(response.body().toString());
+                    Log.i(TAG, "Request enviado." + response.body().toString());
+                } else {
 
+                    Log.i(TAG, "Ocurrió un error.");
+                }
+            }
+
+            @Override
+            public void onFailure(Call<UserPost> call, Throwable t) {
+                Log.e(TAG, "Error al enviar el request.");
+
+            }
+        });
+    }
     public void sendRegister(String name, String surname, String dni, String email, String password) {
         mAPIService.register(ENV, name, surname, Integer.valueOf(dni), email, password, COMMISION, GROUP).enqueue(new Callback<UserPost>() {
             @Override
